Changelist:
- Fixed the bug where occasionally a random unnamed column was added to .csv
- Added support for checking if files exist, are empty and to touch new files safely
- Massively refactored the entire project to simplify the codebase
- Sped up stat gathering times
- Refactored the pipeline into a separate class, to avoid repeated code
- Moved the pipeline print statements into a separate file: pairwise_analysis_print_library.py
- Moved the pipeline itself into a separate file to make it easier to switch between the two game modes
- Removed the need for an extra champ_names file

+++ src/pairwise_analysis_library.py
@@ -29,7 +29,7 @@ class PairwiseChampionData:
 
     def average_placement(self, champion: Champion) -> float:
         placements = self.total_placements(champion)
-        placements = np.array(placements, dtype=np.float)
+        placements = np.array(placements, dtype=np.float32)
         num_placements = np.sum(placements)
         for i in range(1, self.player_count):
             placements[i] *= i + 1
@@ -40,7 +40,7 @@ class PairwiseChampionData:
 
     def __average_placement_with_n(self, champion: Champion) -> (float, int):
         placements = self.total_placements(champion)
-        placements = np.array(placements, dtype=np.float)
+        placements = np.array(placements, dtype=np.float32)
         num_placements = np.sum(placements)
 
         for i in range(1, self.player_count):
